+++ src/features/Todos/pages/AddTodoPage.tsx
@@ -1,16 +1,17 @@
 import { useContext, useRef, type FormEvent } from 'react';
 import { useTranslation } from 'react-i18next';
-import TextAreaField from '../../Formvalidation/TextAreaField';
+import { useNavigate, useParams } from 'react-router-dom';
+
+import type { InputRef } from '../../../types/Reftype';
+import { PriorityType, StatusType } from '../../../types/Todotypes';
 import InputField from '../../Formvalidation/InputField';
 import RadioGroupField from '../../Formvalidation/RadioGroupField';
-import type { InputRef } from '../../../types/Reftype';
-import { useNavigate, useParams } from 'react-router-dom';
+import TextAreaField from '../../Formvalidation/TextAreaField';
 import {
   TodoContext,
   TodoContextActions,
   type Todo,
 } from '../context/TodoContext';
-import { PriorityType, StatusType } from '../../../components/Todo';
 
 function AddTodoPage() {
   const { t } = useTranslation();
