+++ src/features/Todos/pages/TodoDetailPage.tsx
@@ -1,11 +1,12 @@
 import { useContext, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { TodoContext, type Todo } from '../context/TodoContext';
+
 import {
   formatDate,
   getPriorityColor,
   getStatusColor,
 } from '../../../utils/todoUtils';
+import { TodoContext, type Todo } from '../context/TodoContext';
 
 function TodoDetailPage() {
   const { id } = useParams();
@@ -27,7 +28,7 @@ function TodoDetailPage() {
 
     const todo = findTodo(todoContext.state.todoArray, id);
     setTodoData(todo);
-  }, []);
+  },[]);
 
   if (!id) {
     return <h2>Please Provide valid id.</h2>;
