+++ src/types/Todotypes.ts
@@ -1,0 +1,11 @@
+export enum PriorityType {
+  HIGH = 'High',
+  MEDIUM = 'Medium',
+  LOW = 'Low',
+}
+
+export enum StatusType {
+  NOTSELECTED = 'Not Started',
+  INPROGRESS = 'In Progress',
+  COMPLETED = 'Completed',
+}
